// client/src/redux/order/actionType.ts
export enum OrderActionType {
  GET_ORDERS_REQUEST = "GET_ORDERS_REQUEST",
  GET_ORDERS_SUCCESS = "GET_ORDERS_SUCCESS",
  GET_ORDERS_ERROR = "GET_ORDERS_ERROR",

  GET_ORDER_REQUEST = "GET_ORDER_REQUEST",
  GET_ORDER_SUCCESS = "GET_ORDER_SUCCESS",
  GET_ORDER_ERROR = "GET_ORDER_ERROR",

  POST_PAYMENT_PRODUCTS_REQUEST = "POST_PAYMENT_PRODUCTS_REQUEST",
  POST_PAYMENT_PRODUCTS_SUCCESS = "POST_PAYMENT_PRODUCTS_SUCCESS",
  POST_PAYMENT_PRODUCTS_ERROR = "POST_PAYMENT_PRODUCTS_ERROR",

  POST_ORDER_DETAILS_REQUEST = "POST_ORDER_DETAILS_REQUEST",
  POST_ORDER_DETAILS_SUCCESS = "POST_ORDER_DETAILS_SUCCESS",
  POST_ORDER_DETAILS_ERROR = "POST_ORDER_DETAILS_ERROR",
}